fetch: rework to use data sources (#95)
Introduces the concept of a DataSource to Ciel, an abstraction of remote
data repositories. There are two concrete implementations thereof:

* GitHubReleasesDataSource: Which uses GitHub Releases, effectively
behaving identically to Volare.
* StaticWebDataSource: Which uses a static webpage to fetch manifest
files that replace the API calls in questions.

Users can provide the data source over the command-line using
`--data-source {class}:{argument}`, where class can be `github-releases`
or `static-web` respectively. When the data source is `github-releases`,
the argument is in the format '{owner}/{repo}', and when the data source
is a static website, the argument is a URL to the base directory where
the manifest files are located. The manifest files are free to point to
any file on the internet for download information.

The default data source is
`static-web:https://fossi-foundation.github.io/ciel-releases`. To act
identically to Volare, the data source can be
`github-releases:efabless/volare`.